+++ client/src/components/IndicatorsTypes.tsx
@@ -1,7 +1,20 @@
-import { IndicatorsDatasetId } from "./IndicatorsDatasets";
 import { withI18nProps } from "@lingui/react";
 import { withMachineInStateProps } from "state-machine";
 import { AddressPageRoutes } from "routes";
+
+/**
+ * The ids for all datasets shown on the Timeline Tab, _in the order they will appear on the select menu_.
+ *
+ * Note: This separation of array and type definition allows us to more easily iterate over all dataset ids.
+ * See https://stackoverflow.com/a/64174790 for more details on this approach.
+ */
+export const indicatorsDatasetIds = [
+  "hpdcomplaints",
+  "hpdviolations",
+  "dobpermits",
+  "dobviolations",
+] as const;
+export type IndicatorsDatasetId = typeof indicatorsDatasetIds[number];
 
 /**
  * All the time spans you can view data by on the Timeline Tab, _in the order they will appear on the select menu_.
